Completed services
Services are done? I think, unless anything needs to be added for the grand quests

+++ src/main/java/com/group2/Arcazon/repository/OrderRepository.java
@@ -1,5 +1,6 @@
 package com.group2.Arcazon.repository;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -10,5 +11,5 @@ import com.group2.Arcazon.model.Order;
 public interface OrderRepository extends JpaRepository<Order, Long> {
     List<Order> findByCustomer_Id(Long customerId);
     List<Order> findByOrderDateBetween(LocalDateTime startDate, LocalDateTime endDate);
-    List<Order> findByTotalAmountGreaterThan(Double amount);
+    List<Order> findByTotalAmountGreaterThan(BigDecimal amount);
 }
